Working on example

+++ JsonFlow-Example/src/main/java/me/luxoru/example/entity/Entity.java
@@ -8,7 +8,7 @@ public interface Entity {
 
     int getHeight();
 
-    int getMaxHealth();
+    float getMaxHealth();
 
 
 }
